+++ app/Http/Controllers/PostController.php
@@ -109,20 +109,26 @@ class PostController extends Controller
     public function update(UpdatePostRequest $request, Post $post)
     {
         //
-        $validated = $request->validated();
-        if(array_key_exists('image',$validated)) {
-            $image = $validated['image'];
-            $currRelativeUrl = str_replace(url('/'),'',$post->imgPath);
-            Storage::delete($currRelativeUrl);
-            if($image) {
-                $relativeUrl = $image->store('posts');
-                $validated['imgPath'] = url($relativeUrl);
-            }else{
-                $validated['imgPath'] = null;
+        //dd($request->user());
+        if($this->authorize('update',[$post,$request->user()])){
+            $validated = $request->validated();
+            if(array_key_exists('image',$validated)) {
+                $image = $validated['image'];
+                $currRelativeUrl = str_replace(url('/'),'',$post->imgPath);
+                Storage::delete($currRelativeUrl);
+                if($image) {
+                    $relativeUrl = $image->store('posts');
+                    $validated['imgPath'] = url($relativeUrl);
+                }else{
+                    $validated['imgPath'] = null;
+                }
             }
+            $post = $this->postRepository->update($post->id,$validated);
+            return new PostResource($post);
         }
-        $post = $this->postRepository->update($post->id,$validated);
-        return new PostResource($post);
+        return response()->json([
+            'error' => 'Unauthorized'
+        ]);
     }
 
     /**
@@ -130,14 +136,20 @@ class PostController extends Controller
      */
     public function destroy(Request $request, string $id)
     {
-        $deleted = $this->postRepository->delete($id);
-        if (!$deleted) {
-            return response()->json(["message" => "Post not found"], 404);
+        $post = $this->postRepository->find($id);
+        if($this->authorize('delete',[$post, $request->user()])){
+            $deleted = $this->postRepository->delete($id);
+            if (!$deleted) {
+                return response()->json(["message" => "Post not found"], 404);
+            }
+            try {
+                return response()->json(["message" => "Delete post sucessfully"]);
+            } catch (Exception $e) {
+                return response()->json(["message" => "Post not found"], 404);
+            }
         }
-        try {
-            return response()->json(["message" => "Delete post sucessfully"]);
-        } catch (Exception $e) {
-            return response()->json(["message" => "Post not found"], 404);
-        }
+        return response()->json([
+            'error' => 'Unauthorized'
+        ],401);
     }
 }
